Rename run_SherpaCKKWL to run_sherpackkwl for consistency
